\name{Sample1}
\alias{Sample1}
\docType{data}
\title{Sample data 1 to run examples}
\description{Dataset with a sample of size 373 from a population of size 570. It is a dataframe with 373 observations of 8 variables.}
\usage{data(Sample1)}
\examples{
data(Sample1) ###Loads the Sample1 dataset
}
\keyword{datasets}
